added more seeding a readmy and fixed some ui bugs

+++ new-sail-application/app/Http/Controllers/MeetingController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\StoreMeetingRequest;
 use App\Http\Requests\UpdateMeetingRequest;
 use App\Models\Meeting;
+use Carbon\Carbon;
 use Illuminate\Http\Request;
 use Tymon\JWTAuth\Facades\JWTAuth;
 
@@ -15,13 +16,13 @@ class MeetingController extends Controller
      *
      * @return \Illuminate\Http\JsonResponse
      */
-    public function index(Request  $request)
+    public function index(Request $request)
     {
-        $jwt= JWTAuth::parseToken()->authenticate();
+        $jwt = JWTAuth::parseToken()->authenticate();
         $meetings = Meeting::with('attendees');
-        
-        if(!$request->has('all') && $jwt->admin === 1){
-            $meetings->where('organization_id',$jwt->organization_id);
+
+        if (!$request->has('all') || $jwt->admin !== 1) {
+            $meetings->where('organization_id', $jwt->organization_id);
         }
         return response()->json($meetings->get());
     }
@@ -34,13 +35,19 @@ class MeetingController extends Controller
      */
     public function store(StoreMeetingRequest $request)
     {
+        $jwt = JWTAuth::parseToken()->authenticate();
+        if ($jwt->admin === 1) {
+            $organization = $request->organization ? $request->organization : $jwt->organization_id;
+        } else {
+            $organization = $jwt->organization_id;
+        }
         $meeting = Meeting::create([
             'date' => $request->date,
             'start_time' => $request->start_time,
             'end_time' => $request->end_time,
             'title' => $request->title,
-            'agenda' => $request->agenda, 
-            'organization_id' => $request->organization
+            'agenda' => $request->agenda,
+            'organization_id' => $organization
         ]);
 
         $meeting->attendees()->sync($request->attendees);
@@ -56,7 +63,15 @@ class MeetingController extends Controller
      */
     public function show(Meeting $meeting)
     {
-        $meeting->load(['attendees','organization']);
+        $jwt = JWTAuth::parseToken()->authenticate();
+        if ($jwt->admin !== 1 && $jwt->organization_id !== $meeting->organization_id) {
+            return response()->json(['error' => 'Unauthorized'], 401);
+        }
+
+        $meeting->load(['attendees', 'organization']);
+
+        $meeting->start_time = Carbon::parse($meeting->start_time)->format('H:i');
+        $meeting->end_time = Carbon::parse($meeting->end_time)->format('H:i');
 
         return response()->json($meeting);
     }
@@ -70,13 +85,24 @@ class MeetingController extends Controller
      */
     public function update(UpdateMeetingRequest $request, Meeting $meeting)
     {
+        $jwt = JWTAuth::parseToken()->authenticate();
+
+        if ($jwt->admin !== 1 && $jwt->organization_id !== $meeting->organization_id) {
+            return response()->JsonResponse(['error' => 'Unauthorized'], 401);
+        }
+
+        if ($jwt->admin === 1) {
+            $organization = $request->organization ? $request->organization : $jwt->organization_id;
+        } else {
+            $organization = $jwt->organization_id;
+        }
         $meeting->update([
             'date' => $request->date,
             'start_time' => $request->start_time,
             'end_time' => $request->end_time,
             'title' => $request->title,
             'agenda' => $request->agenda,
-            'organization_id' => $request->organization
+            'organization_id' => $organization
         ]);
 
         $meeting->attendees()->sync($request->attendees);
@@ -92,8 +118,14 @@ class MeetingController extends Controller
      */
     public function destroy(Meeting $meeting)
     {
+        $jwt = JWTAuth::parseToken()->authenticate();
+        if ($jwt->admin !== 1 && $jwt->organization_id === $meeting->$jwt->organization_id) {
+            return response()->JsonResponse(['error' => 'Unauthorized'], 401);
+        }
+
         $meeting->delete();
 
         return response()->noContent();
+
     }
 }
